fix: ProductSourceRepository 내 타입 변경

+++ src/main/java/apptive/fin/search/repository/ProductSourceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProductSourceRepository extends JpaRepository<ProductSource, Integer> {
+public interface ProductSourceRepository extends JpaRepository<ProductSource, Long> {
     Optional<ProductSource> findByCode(String code);
 }
